+++ Logging/private/Initialize-LoggingTarget.ps1
@@ -1,21 +1,23 @@
 function Initialize-LoggingTarget {
     param()
 
-    $Targets = Get-ChildItem "$ScriptRoot\targets" -Filter '*.ps1'
-    if ($Logging.CustomTargets) {
-        if (Test-Path $Logging.CustomTargets) {
-            $Targets += Get-ChildItem $Logging.CustomTargets -Filter '*.ps1'
-        }
+    $targets = @()
+    $targets += Get-ChildItem "$ScriptRoot\targets" -Filter '*.ps1'
+
+    if ((![String]::IsNullOrWhiteSpace($Logging.CustomTargets)) -and (Test-Path -Path $Logging.CustomTargets -PathType Container)) {
+        $targets += Get-ChildItem -Path $Logging.CustomTargets -Filter '*.ps1'
     }
 
-    foreach ($Target in $Targets) {
-        $Module = . $Target.FullName
-        $LogTargets[$Module.Name] = @{
-            Init           = $Module.Init
-            Logger         = $Module.Logger
-            Description    = $Module.Description
-            Configuration  = $Module.Configuration
-            ParamsRequired = $Module.Configuration.GetEnumerator() | Where-Object {$_.Value.Required -eq $true} | Select-Object -ExpandProperty Name
+    Write-Verbose -Message ("{0} :: {1} targets configured in sum." -f $MyInvocation.MyCommand,$targets.Length)
+
+    foreach ($target in $targets) {
+        $module = . $target.FullName
+        $Script:LogTargets[$module.Name] = @{
+            Init           = $module.Init
+            Logger         = $module.Logger
+            Description    = $module.Description
+            Configuration  = $module.Configuration
+            ParamsRequired = $module.Configuration.GetEnumerator() | Where-Object {$_.Value.Required -eq $true} | Select-Object -ExpandProperty Name
         }
     }
 }